refactor(composables): moved getters to root folder
