+++ packages/composables/src/index.ts
@@ -30,4 +30,4 @@ export { default as useUserOrder } from './composables/useUserOrder';
 export { default as useUserShipping } from './composables/useUserShipping';
 export { default as useWishlist } from './composables/useWishlist';
 
-export * from './composables/getters';
+export * from './getters';
